Structured the web pages, integrated login with django, created signout, and user auth
I structured the web pages to start with the index.html which if the user is not logged in will go to the login screen. I integrated the login screen with django, I did not integrate the sign up feature. I added a signout link in the index and added a header to welcome the specific user